Fix compare version for CoreDNS
* coredns image doesn't have `v` in it, but Github version does have `v`

[#163345748]

+++ scripts/compare-specs.sh
@@ -21,7 +21,8 @@ names=("kubernetes-dashboard-amd64"
 for i in ${!blob_names[@]}; do
     existing_spec_version=$(bosh blobs --column path | grep "${blob_names[i]}" | grep -o -E 'v?([0-9]+\.)+[0-9]+')
     fetched_spec_version=$(cat "$base_dir/${releases[i]}/tag")
-    if [[ $existing_spec_version != $fetched_spec_version ]]; then
+    if [[ "$existing_spec_version" != "$fetched_spec_version" ]] \
+        && [[ "v$existing_spec_version" != "$fetched_spec_version" ]]; then
         cat <<EOF > "$base_dir/spec-to-update/spec.env"
 export SPEC_RELEASE_DIR=${releases[i]}
 export SPEC_BLOB_NAME=${blob_names[i]}
